requirements.txt
absl-py==2.1.0
aggdraw==1.3.18.post0
appnope==0.1.2
asttokens==2.0.5
astunparse==1.6.3
attrs==23.1.0
Bottleneck==1.3.7
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==0.8.3
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.4.0
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
idna==3.7
iniconfig==1.1.1
ipykernel==6.28.0
ipython==8.25.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
joblib==1.4.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.2
keras==2.12.0
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
ml-dtypes==0.4.0
nbformat==5.9.2
nest-asyncio==1.6.0
numexpr==2.8.7
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
parso==0.8.3
pexpect==4.8.0
pillow==10.4.0
platformdirs==3.10.0
plotly==5.22.0
pluggy==1.0.0
prompt-toolkit==3.0.43
protobuf==4.25.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydot==1.4.1
Pygments==2.15.1
pyparsing==3.0.9
pytest==7.4.4
python-dateutil==2.9.0post0
pytz==2024.1
pyzmq==25.1.2
referencing==0.30.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
stack-data==0.2.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
unicodedata2==15.1.0
urllib3==2.2.2
visualkeras==0.1.3
wcwidth==0.2.5
Werkzeug==3.0.3
wrapt==1.14.1
